pyproject.toml
[project]
name = "pm-pedia-langextract"
version = "0.1.0"
description = "A short description of the project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langextract>=1.0.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.4.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pm_pedia_langextract"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
exclude = [".venv"]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
]
ignore = [
    "G004",
    "F821",
    "C401",
    "C408",
    "RUF001",
    "RUF002",
    "RUF003",
    "ANN101",
    "ANN102",
    "PERF203",
    "PERF401",
    "PLR2004",
    "PLC2401",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "E402", "PLR2004", "PLR0913"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101", "B311", "B404", "B603"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov-branch",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "urllib3>=2.5.0",
]
